+++ src/MyMediaLite/ItemRecommendation/ItemRecommender.cs
@@ -65,7 +65,7 @@ namespace MyMediaLite.ItemRecommendation
 		}
 
 		///
-		public abstract double Predict(int user_id, int item_id);
+		public abstract float Predict(int user_id, int item_id);
 
 		///
 		public virtual bool CanPredict(int user_id, int item_id)
