+++ src/MyMediaLite/RatingPrediction/ITimeAwareRatingPredictor.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2011 Zeno Gantner
+// Copyright (C) 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -36,6 +36,6 @@ namespace MyMediaLite.RatingPrediction
 		/// <param name='user_id'>the user ID</param>
 		/// <param name='item_id'>the item ID</param>
 		/// <param name='time'>the time of the rating event</param>
-		double Predict(int user_id, int item_id, DateTime time);
+		float Predict(int user_id, int item_id, DateTime time);
 	}
 }
