+++ src/MyMediaLite/RatingPrediction/ItemKNN.cs
@@ -52,7 +52,7 @@ namespace MyMediaLite.RatingPrediction
 		/// <param name="user_id">the user ID</param>
 		/// <param name="item_id">the item ID</param>
 		/// <returns>the predicted rating</returns>
-		public override double Predict(int user_id, int item_id)
+		public override float Predict(int user_id, int item_id)
 		{
 			if ((user_id > MaxUserID) || (item_id > correlation.NumberOfRows - 1))
 				return baseline_predictor.Predict(user_id, item_id);
@@ -74,9 +74,9 @@ namespace MyMediaLite.RatingPrediction
 						break;
 				}
 
-			double result = baseline_predictor.Predict(user_id, item_id);
+			float result = baseline_predictor.Predict(user_id, item_id);
 			if (weight_sum != 0)
-				result += sum / weight_sum;
+				result += (float) (sum / weight_sum);
 
 			if (result > MaxRating)
 				result = MaxRating;
